requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
Flask==2.2.3
fonttools==4.39.3
frozendict==2.3.6
fundamentalanalysis==0.2.14
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
multitasking==0.0.11
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4
uritemplate==4.1.1
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
yfinance==0.2.14
